requirements.txt
astroid==1.5.3
boto3==1.4.4
botocore==1.5.56
colorama==0.3.9
django==1.11.1
djangorestframework==3.6.3
docutils==0.13.1
isort==4.2.15
jmespath==0.9.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.1
pynamodb==3.0.1
python-dateutil==2.6.0
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
wrapt==1.10.10
